+++ php/get.php
@@ -87,11 +87,12 @@ switch ($command) {
     case 'search':
         $search = check_key_json('search', $inputJSON, false);
         $random = check_key_json('random', $inputJSON, false);
+        $limit = check_key_json('limit', $inputJSON, false);
 
         if (!$search)
             http_error(400, 'No search provided');
 
-        $result = $db->search($search, $random);
+        $result = $db->search($search, $random, $limit);
 
         http_response(stringifier($result));
         break;
@@ -130,7 +131,8 @@ switch ($command) {
 
 http_message(400, 'Bad request');
 
-
+} catch(HTTPException $e) {
+    http_error($e->getCode(), $e->getMessage());
 } catch(Exception $e) {
-    http_error(500, $e->getMessage());
+    http_error(400, $e->getMessage());
 }