+++ src/index.js
@@ -14,6 +14,12 @@ try {
  */
 
 /**
+ * @typedef {Object} SearchResultOptions
+ * @property {(boolean | number)?} [random] - Random result seed, disabled by default, but can activated with true or a given seed
+ * @property {number?} [limit] - Maximum number of results to return
+ */
+
+/**
  * @typedef {Object} EditFieldOption
  * @property {string | number} id - The affected element
  * @property {string} field - The field to edit
@@ -250,11 +256,38 @@ class Collection {
 	/**
 	 * Search through the collection
 	 * @param {SearchOption[]} options - Array of search options
-	 * @param {boolean | number} [random] - Random result seed, disabled by default, but can activated with true or a given seed
+	 * @param {(boolean|number|SearchResultOptions)?} [resultOptions] - Search result options
 	 * @returns {Promise<T[]>} The found elements
 	 */
-	async search(options, random = false) {
+	async search(options, resultOptions = undefined) {
 		if (!Array.isArray(options)) throw new TypeError("searchOptions shall be an array");
+		if (
+			resultOptions !== undefined &&
+			typeof resultOptions !== "number" &&
+			typeof resultOptions !== "boolean" &&
+			typeof resultOptions !== "object"
+		)
+			throw new TypeError("Incorrect search result options");
+
+		const { random = false, limit = undefined } = resultOptions || {};
+
+		if (
+			limit !== undefined &&
+			(typeof limit !== "number" || limit <= 0 || !Number.isInteger(limit))
+		)
+			throw new TypeError(
+				`${JSON.stringify(limit)} search option limit must be a positive integer`,
+			);
+
+		if (
+			random !== undefined &&
+			random !== false &&
+			random !== true &&
+			(typeof random !== "number" || !Number.isInteger(random))
+		)
+			throw new TypeError(
+				`${JSON.stringify(random)} search option random must be a boolean or an integer`,
+			);
 
 		options.forEach((option) => {
 			if (option.field === undefined || option.criteria === undefined || option.value === undefined)
@@ -273,14 +306,16 @@ class Collection {
 			search: options,
 		};
 
-		if (random !== false) {
+		if (limit !== undefined) {
+			params.limit = limit;
+		}
+
+		if (random !== undefined && random !== false) {
+			params.random = parseInt(random);
 			if (random === true) {
 				params.random = {};
 			} else {
 				const seed = parseInt(random);
-				if (isNaN(seed))
-					throw new TypeError("random takes as parameter true, false or an integer value");
-
 				params.random = { seed };
 			}
 		}